% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/callPatientList.R
\name{callPatientList}
\alias{callPatientList}
\title{callPatientList Function}
\usage{
callPatientList(dayRequest, patientData)
}
\arguments{
\item{day}{is the day in the patient data frame from which to extract the patient requests}
}
\description{
This function generates a daily list of patient requests for a simulation. The
list contains patient objects with defined characteristics and is returned
when the function is called.
}
\examples{
callPatientList(day=1,patientData)
Data Structure:

|  day  |  type  |  team  |  provider  |
   1       acute      1         2
   1       routine    1         1...
}
\keyword{containing}
\keyword{data}
\keyword{demand}
\keyword{extract}
\keyword{frame}
\keyword{from}
\keyword{is}
\keyword{patient}
\keyword{patientData}
\keyword{the}
\keyword{to}
\keyword{which}

