% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/genericFirstAvailableMTF.R
\name{genericFirstAvailableMTF}
\alias{genericFirstAvailableMTF}
\title{genericFirstAvailableMTF Function}
\usage{
genericFirstAvailableMTF(schedule, patient, Day, crossbook = TRUE)
}
\arguments{
\item{schedule}{provides the schedule (list type object) to search for an open appt}

\item{patient}{provides the patient class with the associated attributes}

\item{Day}{identifies the day in the schedule to begin searching for an appt}

\item{crossbook}{indicates if crossbooking to the full clinic is authorized. Default is TRUE.}
}
\description{
This function schedules a patient into the first available appointment
beginning on the specified day. It moves through the hierarchy of the assigned
provider, to the assigned care team, and then to the full clinic if cross-booking
is authorized. The function returns a vector (list) with the position of the
appt to book and whether access and continuity were met. The default is that
access was not met since this function is called when an appt in the access
standard is not available.
}
\examples{
genericFirstAvailable(schedule,patient,Day=5,crossbook=TRUE)
}
\keyword{firstAvailable,}
\keyword{scheduling}

