% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/horizon.R
\name{horizon}
\alias{horizon}
\title{horizon Function}
\usage{
horizon(type, day)
}
\arguments{
\item{type}{provides the appointment type to base the horizon on.}

\item{day}{identifies the day in the schedule to count the horizon from.}
}
\description{
Returns the number of schedule days of the access standard for the appt type
}
\examples{
horizon(type="routine", day = 10)
}
\keyword{accessStandard}
\keyword{horizon,}

