% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/makeMTFSchedule.R
\name{makeMTFSchedule}
\alias{makeMTFSchedule}
\title{makeMTFSchedule Function}
\usage{
makeMTFSchedule(days = 100, teams = 1, provPerTeam = 1, act = 0,
  routine = 0, well = 0, avail = 1)
}
\arguments{
\item{days}{defines the number of days in the schedule. Defaults to 100 days}

\item{teams}{defines the number of teams in the clinic. Defaults to 1 team.}

\item{provPerTeam}{defines the number of providers on each team. Defaults to 1.}

\item{act}{defines the number of appts to reserve for acute patient requests. Defaults to 0.}

\item{routine}{defines the number of appts in the routine appt pool to book. Defaults to 0.}

\item{well}{defines the number of appts in the well appt pool to book. Defaults to 0.}

\item{avail}{defines the percentage of time providers are available in the schedule. Defaults to 1}
}
\description{
This function allows you to create a new MTF schedule to populate with the
three triaged appt types (acute, routine, well).
}
\examples{
makeMTFSchedule(days=100, teams=1, provPerTeam=1, act=5, routine=13, well=1, avail=1)
}
\keyword{schedule}

