% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/makeSchedule.R
\name{makeSchedule}
\alias{makeSchedule}
\title{makeSchedule Function}
\usage{
makeSchedule(days = 100, teams = 1, provPerTeam = 1, act = 0,
  nonact = 0, avail = 1, seed = 5)
}
\arguments{
\item{days}{defines the number of days in the schedule. Defaults to 100 days}

\item{teams}{defines the number of teams in the clinic. Defaults to 1 team.}

\item{provPerTeam}{defines the number of providers on each team. Defaults to 1.}

\item{act}{defines the number of appts to reserve for acute patient requests. Defaults to 0.}

\item{nonact}{defines the number of appts in the general appt pool to book. Defaults to 0.}

\item{avail}{defines the percentage of time providers are available in the schedule. Defaults to 1}
}
\description{
This function allows you to create a new schedule to populate with act and
non-act appts
}
\examples{
makeSchedule(days=100, teams=1, provPerTeam=1, act=5, nonact=13, avail=1)
}
\keyword{schedule}

