% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/patient.R
\name{patient}
\alias{patient}
\title{Patient Function}
\usage{
patient(type = "acute", provider, team)
}
\arguments{
\item{type}{defines the type of appt requested. Defaults to acute.}

\item{provider}{defines the provider ID within the care team to which the patient is assigned}

\item{team}{defines the team ID of the team to which the patient is assigned}
}
\description{
This function allows you to create a Patient object.
}
\examples{
patient(type="routine", provider=2, team=1)
}
\keyword{patient}

