% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PatientGeneration.R
\name{PatientGeneration}
\alias{PatientGeneration}
\title{PatientGeneration Function}
\usage{
PatientGeneration(lambda, teams, provPerTeam, apptRatio, seed = 5)
}
\arguments{
\item{teams}{defines the number of teams in the clinic}

\item{provPerTeam}{defines the number of providers on each team}

\item{apptRatio}{is a named vector with correct element names with the proportion of each appt type. Ex: c(0.3,0.5,0.2)}

\item{lamdba}{the rate provided to the poisson random number generator}
}
\description{
This function generates a daily list of patient requests for a simulation. The
list contains patient objects with defined characteristics and is returned
when the function is called.
}
\examples{
PatientGeneration(lambda,teams,provPerTeam,apptRatio,seed=5)
}
\keyword{PatientGeneration,}
\keyword{patientList,}
\keyword{patients}

